+++ src/CssPosition.ts
@@ -2,13 +2,13 @@
  * Store position information for a node
  */
 export default class Position {
-  start: { line: number; column: number };
-  end: { line: number; column: number };
+  start: { line: number; column: number; offset?: number };
+  end: { line: number; column: number; offset?: number };
   source?: string;
 
   constructor(
-    start: { line: number; column: number },
-    end: { line: number; column: number },
+    start: { line: number; column: number; offset?: number },
+    end: { line: number; column: number; offset?: number },
     source: string,
   ) {
     this.start = start;